backend/requirements.txt
langchain
langchain_openai
ipykernel
streamlit
streamlit-aggrid
streamlit-nested-layout
python-dotenv
flask